+++ public/js/controller/post.js
@@ -1,6 +1,5 @@
 "use strict";
-angular.module('fpm.c.post', ['fpm.service', 'fpm.filter'])
-  .controller('PostCtrl', ['$scope', '$ngFpmcService',
+app.controller('PostCtrl', ['$scope', '$ngFpmcService',
     function ($scope, $ngFpmcService) {
         const { Func, Query } = $ngFpmcService;
         new Query('usr_userinfo')
@@ -9,13 +8,14 @@ angular.module('fpm.c.post', ['fpm.service', 'fpm.filter'])
           .then(console.log)
           .catch(console.error)
     }])
-  .controller('PostCreateCtrl', ['$scope', '$ngFpmcService', '$q', '$http',
-    function ($scope, $ngFpmcService, $q, $http) {
+  .controller('PostCreateCtrl', ['$scope', '$ngFpmcService', 'kit', '$http',
+    function ($scope, $ngFpmcService, kit, $http) {
         const { Func, Query } = $ngFpmcService;
         $scope.post = {
           content: '<p>AAABC</p>',
           category: { id: 1},
           title: '',
+          cover: 'http://placehold.it/150x100',
         };
         // load the tags /cates
         new Query('cms_tags')
@@ -36,55 +36,38 @@ angular.module('fpm.c.post', ['fpm.service', 'fpm.filter'])
           })
           .catch(console.error);
 
-        ClassicEditor
-            .create( document.querySelector( '#editor' ), {
-              ckfinder: {
-                uploadUrl: '/upload',
-              }
-            } )
-            .then( editor => {
-              $scope.editor = editor;
+        if(ClassicEditor){
+          ClassicEditor
+              .create( document.querySelector( '#editor' ), {
+                ckfinder: {
+                  uploadUrl: '/upload',
+                }
+              } )
+              .then( editor => {
+                $scope.editor = editor;
+              })
+              .catch( error => {
+                  console.error( error );
+              } );
+        }
+        $scope.imageUpload = function(){
+
+          kit.chooseFile().then(function(rsp){
+            console.log(rsp.data);
+            $scope.$apply(function(){
+              $scope.post.cover = rsp.data.url;
             })
-            .catch( error => {
-                console.error( error );
-            } );
-
+          })
+        }
         $scope.save = function(){
-          // var fd = new FormData(); //初始化一个FormData实例
-          // fd.append('upload', document.querySelector('#upload').files[0]);
-
-          // $http({
-          //   method:'POST',
-          //   url: '/upload',
-          //   headers: {
-          //     'Content-Type': undefined
-          //   },
-          //   transformRequest: angular.identity,
-          //   data: fd,
-          // })
-          // .then((rsp) => {
-          //   console.log(rsp.data);
-          // })
-          // .catch((err) => {
-          //   Swal({
-          //     position: 'center',
-          //     type: 'error',
-          //     title: '上传失败',
-          //     showConfirmButton: false,
-          //     timer: 1500
-          //   })
-          // });
-          // Swal({
-          //   position: 'center',
-          //   type: 'success',
-          //   title: '保存成功',
-          //   showConfirmButton: false,
-          //   timer: 1500
-          // }).then(() => {
-          //   // alert(1)
-          // });
-          // $scope.post.content = $contentDom.val();
+          kit.toast('hi')
+            .then(function(){
+              kit.confirm('Yes or no?')
+                .then(function(result){
+                  kit.logger.debug(result.value, 'POST.js');
+                })
+                .catch(kit.logger.error);
+            });
           $scope.post.content = $scope.editor.getData();
-          console.log($scope.post);
         }
     }])